+++ RELEASE_NOTES.md
@@ -3,6 +3,7 @@ Release Notes
 
 ## 1.9.1
 * Managed Clusters (AKS): Support for workload identity, OIDC issuer, image cleaner, and Defender.
+* Managed Clusters (AKS): Default to use MSI for the service principal profile to align with CLI and Portal.
 * User Assigned Identities: Support for `depends_on`.
 
 ## 1.9.0